[Store] Organize batch OpLog snapshot metadata
