+++ mooncake-store/include/ha/snapshot/batch_oplog/metadata.h
