+++ mooncake-store/src/ha/snapshot/batch_oplog/metadata.cpp
@@ -1,4 +1,4 @@
-#include "ha/snapshot/batch_oplog_snapshot_types.h"
+#include "ha/snapshot/batch_oplog/metadata.h"
 
 #include <charconv>
 #include <memory>
